Fix SLAM3R camera pose extraction and TSDF integration
- Implement proper SVD-based pose extraction in streaming_slam3r.py to replace hardcoded identity matrix
- Fix initialization bug where pts3d_cam and pts3d_world were incorrectly set to same values
- Add comprehensive debugging for camera pose through entire pipeline
- Update docker-compose.yml TSDF bounds to match actual scene coordinates
- Add robustness improvements: batch size validation, SVD error handling, degenerate case detection
- Refactor frame_processor for better modularity and add Rerun visualization support
- Add debugging output in mesh_service shared memory and TSDF integration

This fixes the issue where all camera poses were identity matrices, causing incorrect mesh reconstruction.

+++ mesh_service/src/normal_providers/camera_based_normal_provider.cu
@@ -45,8 +45,8 @@ __global__ void computeCameraBasedNormalsKernel(
     
     // Debug output for first few normals
     if (idx < debug_normal_count) {
-        printf("[CAMERA NORMAL] Point %d: pos=[%.3f,%.3f,%.3f], normal=[%.3f,%.3f,%.3f], dist=%.3f\n",
-               idx, point.x, point.y, point.z, 
+        printf("[CAMERA NORMAL] Point %lu: pos=[%.3f,%.3f,%.3f], normal=[%.3f,%.3f,%.3f], dist=%.3f\n",
+               (unsigned long)idx, point.x, point.y, point.z, 
                normals[idx].x, normals[idx].y, normals[idx].z, dist);
     }
 }
